Skeleton refactor: Put drawBackdrop rect in the Skeleton class
