+++ packages/core/src/components/graphs/skeleton-vert-or-horiz.ts
@@ -7,7 +7,6 @@ import { axisLeft, axisBottom } from "d3-axis";
 
 export class SkeletonVertOrHoriz extends Skeleton {
 	type = "skeleton";
-	backdrop: any;
 	xScale: any;
 	yScale: any;
 
@@ -34,26 +33,6 @@ export class SkeletonVertOrHoriz extends Skeleton {
 			this.drawXGrid(animate);
 		}
 		this.setStyle();
-	}
-
-	drawBackdrop() {
-		const svg = this.parent;
-
-		this.backdrop = DOMUtils.appendOrSelect(svg, "svg.chart-skeleton-backdrop");
-		const backdropRect = DOMUtils.appendOrSelect(this.backdrop, "rect.chart-skeleton-backdrop");
-		backdropRect
-			.attr("width", "100%")
-			.attr("height", "100%");
-
-		const [xScaleStart, xScaleEnd] = this.xScale.range();
-		const [yScaleEnd, yScaleStart] = this.yScale.range();
-
-		this.backdrop
-			.merge(backdropRect)
-			.attr("x", xScaleStart)
-			.attr("y", yScaleStart)
-			.attr("width", xScaleEnd - xScaleStart)
-			.attr("height", yScaleEnd - yScaleStart);
 	}
 
 	drawXGrid(animate: boolean) {
